pyproject.toml
[project]
name = "finance-wiz"
version = "0.1.0"
description = "Local stock prediction engine with a swappable-algorithm pipeline."
readme = "README.md"
authors = [
    { name = "jcnaef", email = "jcnaef@gmail.com" }
]
requires-python = ">=3.11,<3.13"
dependencies = [
    "lightgbm>=4.6.0",
    "pandas>=3.0.2",
    "pyarrow>=24.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "skorch>=1.3.1",
    "torch>=2.11.0",
    "typer>=0.25.0",
    "xgboost>=3.2.0",
    "yfinance>=1.3.0",
]

[project.scripts]
fwiz = "finance_wiz.cli:app"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
